Added an implementation of GameTemplate.Output and GameTemplate.Input that is created with Swing. This implementation is to be used to display the user's hand.
When given the string representation of a hand, it is able to display all the cards in the hand as buttons on the screen.

When used as an input, it uses popups to prompt the user for input or waits for the user to press a button before moving on.